Add editorconfig and fix warnings
