+++ Directory.build.props
@@ -1,0 +1,27 @@
+﻿<Project>
+  <PropertyGroup>
+    <Authors>Nathan Le Sommer</Authors>
+    <Copyright>Copyright 2020</Copyright>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <!-- Warnings as errors -->
+  <PropertyGroup Condition=" '$(Configuration)'=='Release' ">
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.354">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.354">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <AdditionalFiles Include="$(MSBuildThisFileDirectory)stylecop.json" Link="Properties/%(Filename)%(Extension)" />
+  </ItemGroup>
+</Project>
