--- src/IvorySaga.Api/GlobalSuppressions.cs
@@ -1,9 +1,0 @@
-﻿// This file is used by Code Analysis to maintain SuppressMessage
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given
-// a specific target and scoped to a namespace, type, member, etc.
-
-using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "No Async suffix on controller endpoint methods.", Scope = "type", Target = "~T:IvorySaga.Api.Controllers.SagasController")]
-[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "No Async suffix on controller endpoint methods.", Scope = "type", Target = "~T:IvorySaga.Api.Controllers.ChaptersController")]
