+++ src/IvorySaga.Api/Models/SagaModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 
 namespace IvorySaga.Api.Models
 {
