+++ src/IvorySaga/(Exceptions)/ChapterNotFoundException.cs
@@ -5,12 +5,13 @@ namespace IvorySaga
     [Serializable]
     public class ChapterNotFoundException : Exception
     {
-        public ChapterNotFoundException() { }
+        public ChapterNotFoundException()
+        {
+        }
 
         public ChapterNotFoundException(string sagaId, string chapterId)
             : base(string.Format("Chapter {0} not found for saga {1}.", sagaId, chapterId))
         {
-
         }
     }
 }
