+++ src/IvorySaga/(Exceptions)/SagaNotFoundException.cs
@@ -5,12 +5,13 @@ namespace IvorySaga
     [Serializable]
     public class SagaNotFoundException : Exception
     {
-        public SagaNotFoundException() { }
+        public SagaNotFoundException()
+        {
+        }
 
         public SagaNotFoundException(string id)
             : base(string.Format("Saga not found: {0}", id))
         {
-
         }
     }
 }
