+++ src/IvorySaga/Data/Chapter.cs
@@ -1,6 +1,4 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
-using System;
+﻿using System;
 
 namespace IvorySaga.Data
 {
@@ -9,7 +7,6 @@ namespace IvorySaga.Data
     /// </summary>
     public sealed class Chapter
     {
-        //[BsonId]
         public Guid Id { get; set; } = default!;
 
         public Guid SagaId { get; set; } = default!;
