+++ src/IvorySaga/Data/Saga.cs
@@ -1,6 +1,4 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
-using System;
+﻿using System;
 
 namespace IvorySaga.Data
 {
