+++ src/IvorySaga/Mongo/IIvorySagaDatabaseSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace IvorySaga.Mongo
+{
+    public interface IIvorySagaDatabaseSettings
+    {
+        string SagasCollectionName { get; set; }
+
+        string ChaptersCollectionName { get; set; }
+
+        string ConnectionString { get; set; }
+
+        string DatabaseName { get; set; }
+    }
+}
