+++ src/IvorySaga/Queries/GetChapterQuery.cs
@@ -1,22 +1,23 @@
-﻿using IvorySaga.Data;
-using IvorySaga.Services;
-using MediatR;
-using System;
+﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using IvorySaga.Data;
+using IvorySaga.Services;
+using MediatR;
 
 namespace IvorySaga.Queries
 {
     public sealed class GetChapterQuery : IRequest<Chapter>
     {
-        public Guid SagaId { get; } = default!;
-        public Guid ChapterId { get; } = default!;
-
         public GetChapterQuery(Guid sagaId, Guid chapterId)
         {
             SagaId = sagaId;
             ChapterId = chapterId;
         }
+
+        public Guid SagaId { get; } = default!;
+
+        public Guid ChapterId { get; } = default!;
 
         internal sealed class Handler : IRequestHandler<GetChapterQuery, Chapter>
         {
