+++ src/IvorySaga/Queries/GetChaptersQuery.cs
@@ -1,22 +1,22 @@
-﻿using IvorySaga.Data;
-using IvorySaga.Services;
-using MediatR;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using IvorySaga.Data;
+using IvorySaga.Services;
+using MediatR;
 
 namespace IvorySaga.Queries
 {
     public sealed class GetChaptersQuery : IRequest<IEnumerable<Chapter>>
     {
-        public Guid SagaId { get; } = default!;
-
         public GetChaptersQuery(Guid id)
         {
             SagaId = id;
         }
+
+        public Guid SagaId { get; } = default!;
 
         internal sealed class Handler : IRequestHandler<GetChaptersQuery, IEnumerable<Chapter>>
         {
