+++ src/IvorySaga/Services/ChapterRepository.cs
@@ -1,11 +1,11 @@
-﻿using IvorySaga.Data;
-using IvorySaga.Mongo;
-using MongoDB.Driver;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using IvorySaga.Data;
+using IvorySaga.Mongo;
+using MongoDB.Driver;
 
 namespace IvorySaga.Services
 {
